Priority Flag Icons and Styles
Added priority flag icons before the description of each task as well as CSS styling for icon colors.

This takes into consideration only 3 priority flags.  If the task is left as priority 1, which I believe is the default, no flag will show, otherwise 2, 3, and 4 will produce flags.

Todoist priority 1 tasks are shown in the data as "4" and the lowest priority tasks are shown as "1".  It is all in reverse order from what you would consider high priority (i.e. Priority 1.)